pyproject.toml
[tool.poetry]
name = "game_of_life"
version = "0.1.0"
description = ""
authors = ["Martin Kuchynka <kuchynkm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.5.3"
numpy = "1.20"
Pillow = "^8.3.2"
scipy = "~1.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
